// src/renderer/src/utils/modules/FileValidate/enum.ts
/**
 * 文件验证错误类型
 */
export enum FileValidationError {
  SIZE_TOO_LARGE = 'SIZE_TOO_LARGE',
  SIZE_TOO_SMALL = 'SIZE_TOO_SMALL',
  INVALID_TYPE = 'INVALID_TYPE',
  INVALID_EXTENSION = 'INVALID_EXTENSION',
  INVALID_DIMENSION = 'INVALID_DIMENSION',
  CUSTOM_ERROR = 'CUSTOM_ERROR',
  CORRUPTED_FILE = 'CORRUPTED_FILE',
  MAGIC_NUMBER_MISMATCH = 'MAGIC_NUMBER_MISMATCH'
}